backend/requirements.txt
azure-functions>=1.18.0
azure-data-tables>=12.4.0
python-dateutil>=2.8.2
certifi>=2023.7.22

